Update README.md
Generally, we can get the minimum weighted result by finding the smallest two number, plus them, then put the result back to the heap. In this way we can get a Huffman Tree, and know the weight of each number. However, we can have other thoughts, eg., in this code, we get the minimum weighted path by summing the values of all the nodes, and its size is much smaller